完成了班级管理、学员管理、学员信息统计、功能完善

+++ tlias-web-management/src/main/java/com/itheima/service/impl/DeptServiceImpl.java
@@ -1,7 +1,10 @@
 package com.itheima.service.impl;
 
+import com.itheima.exception.BusinessException;
 import com.itheima.mapper.DeptMapper;
+import com.itheima.mapper.EmpMapper;
 import com.itheima.pojo.Dept;
+import com.itheima.pojo.Emp;
 import com.itheima.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +16,9 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
+
     @Override
     public List<Dept> findAll() {
         return deptMapper.findAll();
@@ -20,7 +26,13 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public void deleteById(Integer id) {
-        deptMapper.deleteById(id);
+        List<Emp> empList = empMapper.isHasEmp(id);
+        if(empList.isEmpty()){
+            deptMapper.deleteById(id);
+        }else{
+            throw new BusinessException("该部门有员工，不能删除");
+        }
+
     }
 
     @Override
@@ -41,4 +53,6 @@ public class DeptServiceImpl implements DeptService {
         dept.setUpdateTime(LocalDateTime.now());
         deptMapper.update(dept);
     }
+
+
 }
